requirements.txt
streamlit==1.13.0
numpy==1.23.3
PyYAML==6.0
transformers==4.25.1
sklearn==0.0
scikit-learn==1.1.2
torch==1.13.1
tqdm==4.64.1
docx==0.2.4
# dill нужно скачать
pypdf2==2.12.1
requests==2.28.1